+++ src/main/java/com/antkorwin/springtestmongo/internal/expect/graph/Graph.java
@@ -1,4 +1,4 @@
-package com.antkorwin.springtestmongo.internal.expect;
+package com.antkorwin.springtestmongo.internal.expect.graph;
 
 import java.util.Map;
