+++ src/main/java/com/antkorwin/springtestmongo/internal/expect/graph/IndexedGraph.java
@@ -1,4 +1,4 @@
-package com.antkorwin.springtestmongo.internal.expect;
+package com.antkorwin.springtestmongo.internal.expect.graph;
 
 import java.util.HashSet;
 import java.util.Map;
