+++ src/test/java/com/antkorwin/springtestmongo/internal/TestData.java
@@ -3,6 +3,9 @@ package com.antkorwin.springtestmongo.internal;
 import java.util.List;
 import java.util.Map;
 
+import com.antkorwin.springtestmongo.internal.importdata.ImportFile;
+import com.antkorwin.springtestmongo.internal.importdata.JsonImport;
+
 /**
  * Created on 15.12.2018.
  *
