+++ src/test/java/com/antkorwin/springtestmongo/internal/exportdata/JsonExportTest.java
@@ -1,6 +1,9 @@
-package com.antkorwin.springtestmongo.internal;
+package com.antkorwin.springtestmongo.internal.exportdata;
 
 import com.antkorwin.springtestmongo.Bar;
+import com.antkorwin.springtestmongo.internal.DataSet;
+import com.antkorwin.springtestmongo.internal.exportdata.JsonExport;
+import com.antkorwin.springtestmongo.internal.importdata.ImportFile;
 import org.junit.jupiter.api.Test;
 import org.testcontainers.shaded.com.fasterxml.jackson.databind.ObjectMapper;
 import org.testcontainers.shaded.com.google.common.collect.ImmutableMap;
