+++ src/test/java/com/antkorwin/springtestmongo/internal/importdata/MongoDataImportTest.java
@@ -1,6 +1,8 @@
-package com.antkorwin.springtestmongo.internal;
+package com.antkorwin.springtestmongo.internal.importdata;
 
 import com.antkorwin.springtestmongo.Bar;
+import com.antkorwin.springtestmongo.internal.DataSet;
+import com.antkorwin.springtestmongo.internal.importdata.MongoDataImport;
 import org.assertj.core.groups.Tuple;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
